Refactor: fetcher 위치 변경
constant -> utils 폴더로 이동
