+++ src/features/user/myInvestment/services/myInvestment.service.ts
@@ -2,7 +2,7 @@ import type {
   HoldingRatio,
   UserInvestmentSummary,
 } from "@/entities/user/user.entity";
-import { API_END_POINT } from "@/shared/constants/fetcher";
+import { API_END_POINT } from "@/shared/utils/fetcher";
 
 export async function getInvestmentSummary(): Promise<UserInvestmentSummary> {
   const userId = 1;
